[chore] move collectdreceiver shared code to an internal package
