+++ internal/collectd/go.mod
@@ -1,0 +1,3 @@
+module github.com/open-telemetry/opentelemetry-collector-contrib/internal/collectd
+
+go 1.20
