+++ receiver/collectdreceiver/collectd.go
@@ -6,13 +6,13 @@ package collectdreceiver // import "github.com/open-telemetry/opentelemetry-coll
 import (
 	"encoding/json"
 	"fmt"
-	"strings"
 	"time"
 
 	"go.opentelemetry.io/collector/pdata/pcommon"
 	"go.opentelemetry.io/collector/pdata/pmetric"
 	"go.uber.org/zap"
 
+	"github.com/open-telemetry/opentelemetry-collector-contrib/internal/collectd"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/internal/common/sanitize"
 )
 
@@ -187,7 +187,7 @@ func (cdr *collectDRecord) pointTypeInstance(attrs map[string]string, parts []by
 		return parts
 	}
 
-	instanceName, extractedAttrs := LabelsFromName(cdr.TypeInstance)
+	instanceName, extractedAttrs := collectd.LabelsFromName(cdr.TypeInstance)
 	if instanceName != "" {
 		if len(parts) > 0 {
 			parts = append(parts, '.')
@@ -201,50 +201,6 @@ func (cdr *collectDRecord) pointTypeInstance(attrs map[string]string, parts []by
 		}
 	}
 	return parts
-}
-
-// LabelsFromName tries to pull out dimensions out of name in the format
-// "name[k=v,f=x]-more_name".
-// For the example above it would return "name-more_name" and extract dimensions
-// (k,v) and (f,x).
-// If something unexpected is encountered it returns the original metric name.
-//
-// The code tries to avoid allocation by using local slices and avoiding calls
-// to functions like strings.Slice.
-func LabelsFromName(val *string) (metricName string, labels map[string]string) {
-	metricName = *val
-	index := strings.Index(*val, "[")
-	if index > -1 {
-		left := (*val)[:index]
-		rest := (*val)[index+1:]
-		index = strings.Index(rest, "]")
-		if index > -1 {
-			working := make(map[string]string)
-			dimensions := rest[:index]
-			rest = rest[index+1:]
-			cindex := strings.Index(dimensions, ",")
-			prev := 0
-			for {
-				if cindex < prev {
-					cindex = len(dimensions)
-				}
-				piece := dimensions[prev:cindex]
-				tindex := strings.Index(piece, "=")
-				if tindex == -1 || strings.Contains(piece[tindex+1:], "=") {
-					return
-				}
-				working[piece[:tindex]] = piece[tindex+1:]
-				if cindex == len(dimensions) {
-					break
-				}
-				prev = cindex + 1
-				cindex = strings.Index(dimensions[prev:], ",") + prev
-			}
-			labels = working
-			metricName = left + rest
-		}
-	}
-	return
 }
 
 func isNilOrEmpty(str *string) bool {
@@ -263,7 +219,7 @@ func parseAndAddLabels(labels map[string]string, pluginInstance *string, host *s
 }
 
 func parseNameForLabels(labels map[string]string, key string, val *string) {
-	instanceName, toAddDims := LabelsFromName(val)
+	instanceName, toAddDims := collectd.LabelsFromName(val)
 
 	for k, v := range toAddDims {
 		if _, exists := labels[k]; !exists {
